MRBV material color bug fix

+++ src/underactuated/meshcat_rigid_body_visualizer.py
@@ -130,9 +130,9 @@ class MeshcatRigidBodyVisualizer(LeafSystem):
                         list of >= 3 elements will work), where each element
                         is specified on range [0., 1.], into the equivalent
                         24-bit value 0xRRGGBB. '''
-                        val = 0.
+                        val = 0
                         for i in range(3):
-                            val += (256**(2 - i)) * (255. * rgb[i])
+                            val += (256**(2 - i)) * int(255 * rgb[i])
                         return val
                     rgba = [1., 0.7, 0., 1.]
                     if not self.draw_collision:
